feat: categories page now works

+++ src/main/java/hu/okrim/productratingapp/repository/CategoryRepository.java
@@ -1,8 +1,7 @@
 package hu.okrim.productratingapp.repository;
 
 import hu.okrim.productratingapp.entity.Category;
-import hu.okrim.productratingapp.entity.Person;
 import org.springframework.data.repository.CrudRepository;
 
-public interface CategoryRepository extends CrudRepository<Category, Long> {
+public interface CategoryRepository extends CrudRepository<Category, Integer> {
 }
